feat(endo/init): updated codebase with _document.js.
* Information on using next.js' beforeInteractive strategy [nextjs.org/docs: no beforeInteractive outside docoument](https://nextjs.org/docs/messages/no-before-interactive-script-outside-document)

* tsconfig.json - resolved typescript compilation error for non-relative paths [Resolving TypeScript Errors in Next.js](https://peterkellner.net/2023/09/15/Resolving-TypeScript-Errors-in-Nextjs/)
