+++ components/root/Layout.tsx
@@ -1,0 +1,67 @@
+import '@interchain-ui/react/styles';
+
+import { SignerOptions, wallets } from 'cosmos-kit';
+import { ChainProvider } from '@cosmos-kit/react';
+import { assets, chains } from 'chain-registry';
+import {
+    Box,
+    ThemeProvider,
+    useColorModeValue,
+    useTheme,
+} from '@interchain-ui/react';
+import Head from 'next/head';
+
+const Layout = ({
+    children,
+}: {
+    children: React.ReactNode
+}) => {
+    const { themeClass } = useTheme();
+
+    const signerOptions: SignerOptions = {
+        // signingStargate: () => {
+        //   return getSigningCosmosClientOptions();
+        // }
+    };
+
+    return (
+        <>
+            <Head>
+                <title>Create Cosmos App</title>
+                <meta name="description" content="Generated by create cosmos app" />
+                <link rel="icon" href="/favicon.ico" />
+            </Head> 
+            <ThemeProvider>
+                <ChainProvider
+                    chains={chains}
+                    assetLists={assets}
+                    wallets={wallets}
+                    walletConnectOptions={{
+                        signClient: {
+                            projectId: 'a8510432ebb71e6948cfd6cde54b70f7',
+                            relayUrl: 'wss://relay.walletconnect.org',
+                            metadata: {
+                                name: 'CosmosKit Template',
+                                description: 'CosmosKit dapp template',
+                                url: 'https://docs.cosmology.zone/cosmos-kit/',
+                                icons: [],
+                            },
+                        },
+                    }}
+                    // @ts-ignore
+                    signerOptions={signerOptions}
+                >
+                    <Box
+                        className={themeClass}
+                        minHeight="100dvh"
+                        backgroundColor={useColorModeValue('$white', '$background')}
+                    >
+
+                        {children}
+                    </Box>
+                </ChainProvider>
+            </ThemeProvider>
+        </>);
+}
+
+export default Layout
